moved CommentSection code to separate component
